refactor(frontend): move files into folders
