+++ cmd/gomander/frontend/src/screens/ExperimentalLogsScreen/hooks/useTerminalSearch.ts
